+++ jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  moduleFileExtensions: ['js', 'ts'],
+
+  transform: {
+    '^.+\\.ts$': 'ts-jest',
+  },
+
+  transformIgnorePatterns: ['/node_modules/'],
+
+};